+++ layrzsdk/entities/__init__.py
@@ -1,16 +1,20 @@
 """ Init file """
-# Asset entities
-from .asset import Asset, Sensor, AssetOperationMode
+# Basic entities
+from .asset import Asset
+from .sensor import Sensor
+from .asset_operation_mode import AssetOperationMode
 from .device import Device
 from .custom_field import CustomField
-
-# Messages, Events
-from .message import Position, Message
-from .events import Event, Trigger
-from .cases import User, Comment, Case
-from .checkpoints import Geofence, Waypoint, Checkpoint
+from .message import Message
+from .position import Position
+from .event import Event
+from .trigger import Trigger
+from .case import Case
+from .user import User
+from .comment import Comment
+from .checkpoint import Checkpoint
+from .geofence import Geofence
+from .waypoint import Waypoint
 
 # Charta entities
-from .range_charts import LineChart, AreaChart, ColumnChart, BarChart
-from .charts import ChartDataSerie, ChartAlignment, ChartException, ChartConfiguration, ChartDataType
-from .instant_charts import PieChart, RadialBarChart
+from .charts import *